Workaround "__udivdi3" error. 64-bit division isn't natively supported in x86. The workaround is to explicitly link the 64-bit routines from libgcc.
The Intel SDM states that for the DIV and IDIV instructions 64-bit quotients are not supported in i386. In i386-elf-gcc this is `long long`. Attempts to divide a long long generate errors like "__umoddi3 __udivdi3 undefined". The workaround is to implement these yourself, with 200 lines of code, see [here(http://fxr.watson.org/fxr/source/libkern/divdi3.c?v=DFBSD). Or, to link the gcc implementation as described [here](https://stackoverflow.com/questions/34532528/undefined-reference-to-divdi3).